requirements.txt
tensorflow==2.15.0
numpy==1.26.2
matplotlib==3.8.2
pillow==10.1.0
networkx==3.1